+++ cookbooks/wordpress-kibo-training/recipes/app.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: wordpress
+# Cookbook Name:: wordpress-kibo-training
 # Recipe:: app
 #
 # Copyright 2009-2010, Opscode, Inc.
@@ -17,7 +17,7 @@
 # limitations under the License.
 #
 
-include_recipe "wordpress::database"
+include_recipe "wordpress-kibo-training::database"
 
 ::Chef::Recipe.send(:include, Opscode::OpenSSL::Password)
 node.set_unless['wordpress']['keys']['auth'] = secure_password