+++ cookbooks/wordpress-kibo-training/recipes/database.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: wordpress
+# Cookbook Name:: wordpress-kibo-training
 # Recipe:: database
 # Author:: Lucas Hansen (<lucash@opscode.com>)
 # Author:: Julian C. Dunn (<jdunn@getchef.com>)
@@ -31,7 +31,7 @@ mysql2_chef_gem 'default' do
 end
 
 ::Chef::Recipe.send(:include, Opscode::OpenSSL::Password)
-::Chef::Recipe.send(:include, Wordpress::Helpers)
+::Chef::Recipe.send(:include, wordpress-kibo-training::Helpers)
 
 node.set_unless['wordpress']['db']['pass'] = secure_password
 node.save unless Chef::Config[:solo]