+++ cookbooks/wordpress-kibo-training/CHANGELOG.md
@@ -1,5 +1,15 @@
 CHANGELOG
 
+Version 1.0.2
+-modified:   recipes/apache.rb
+-modified:   recipes/app.rb
+-modified:   recipes/database.rb
+-modified:   recipes/default.rb
+-modified:   recipes/languages.rb
+-modified:   recipes/nginx.rb
+
+
+
 Version 1.0.1
 -Changed attributes for db root password and db instance name
 -Changed max file size upload to 250M